refactor: Renamed routes endpoints
