+++ .github/labeler.yml
@@ -13,44 +13,50 @@
 'module: database':
 - src/app/db/*
 
+'module: schemas':
+- src/app/schemas/*
+
+'module: models':
+- src/app/models/*
+
 'module: services':
 - src/app/services/*
 
 'route: accesses':
-- src/app/api/routes/accesses.py
+- src/app/api/endpoints/accesses.py
 - src/app/api/crud/accesses.py
 
 'route: alerts':
-- src/app/api/routes/alerts.py
+- src/app/api/endpoints/alerts.py
 - src/app/api/crud/alerts.py
 
 'route: devices':
-- src/app/api/routes/devices.py
+- src/app/api/endpoints/devices.py
 
 'route: events':
-- src/app/api/routes/events.py
+- src/app/api/endpoints/events.py
 
 'route: groups':
-- src/app/api/routes/groups.py
+- src/app/api/endpoints/groups.py
 - src/app/api/crud/groups.py
 
 'route: installations':
-- src/app/api/routes/installations.py
+- src/app/api/endpoints/installations.py
 
 'route: login':
-- src/app/api/routes/login.py
+- src/app/api/endpoints/login.py
 
 'route: media':
-- src/app/api/routes/media.py
+- src/app/api/endpoints/media.py
 
 'route: sites':
-- src/app/api/routes/sites.py
+- src/app/api/endpoints/sites.py
 
 'route: users':
-- src/app/api/routes/users.py
+- src/app/api/endpoints/users.py
 
 'route: webhooks':
-- src/app/api/routes/webhooks.py
+- src/app/api/endpoints/webhooks.py
 - src/app/api/crud/webhooks.py
 
 'topic: build':
