+++ src/app/api/endpoints/devices.py
