+++ src/app/api/endpoints/groups.py
