+++ src/app/api/endpoints/installations.py
